file save edge case handled for existence of file

+++ controllers/users_controller.js
@@ -41,7 +41,11 @@ module.exports.update = async function(req, res){
                 if(req.file){
 
                     if(user.avatar){
-                        fs.unlinkSync(path.join(__dirname, '..', user.avatar));
+                        if(fs.existsSync(path.join(__dirname, '..', user.avatar)))
+                        {
+                            fs.unlinkSync(path.join(__dirname, '..', user.avatar));
+                        }
+                        
                     }
 
                     //this is saving the path of uploaded file into the avatar field in the user
